% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos.R
\docType{data}
\name{codigos_estados}
\alias{codigos_estados}
\title{Códigos e nomes dos estados brasileiros (2022)}
\format{
Um data frame com 27 linhas e 3 colunas:
\itemize{
\item \code{codigo_estado} - código do estado;
\item \code{nome_estado} - nome do estado;
\item \code{abrev_estado} - abreviação do nome do estado.
}
}
\source{
\url{https://www.ibge.gov.br/explica/codigos-dos-municipios.php}
}
\usage{
codigos_estados
}
\description{
Uma relação entre os códigos e nomes dos estados brasileiros. Os códigos
foram convertidos para caracteres; os nomes foram convertidos para caracteres
ASCII em caixa alta.
}
\seealso{
\link{codigos_municipios}
}
\keyword{datasets}
