% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos.R
\docType{data}
\name{codigos_municipios}
\alias{codigos_municipios}
\title{Códigos e nomes dos municípios brasileiros (2022)}
\format{
Um data frame com 5570 linhas e 3 colunas:
\itemize{
\item \code{codigo_estado} - código do estado em que o município está localizado;
\item \code{codigo_muni} - código do município;
\item \code{nome_muni} - nome do município.
}
}
\source{
\url{https://www.ibge.gov.br/explica/codigos-dos-municipios.php}
}
\usage{
codigos_municipios
}
\description{
Uma relação entre os códigos e nomes dos municípios brasileiros. Os códigos
foram convertidos para caracteres; os nomes foram convertidos para caracteres
ASCII em caixa alta.
}
\seealso{
\link{codigos_estados}
}
\keyword{datasets}
