% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enderecopadrao.R
\docType{package}
\name{enderecopadrao}
\alias{enderecopadrao}
\alias{enderecopadrao-package}
\alias{pacote-enderecopadrao}
\title{enderecopadrao: Padronizador de Endereços Brasileiros (Brazilian Addresses Standardizer)}
\description{
Padroniza endereços brasileiros a partir de diferentes critérios. Os métodos de padronização incluem apenas manipulações básicas de strings, não oferecendo suporte a correspondências probabilísticas entre strings. (Standardizes brazilian addresses using different criteria. Standardization methods include only basic string manipulation, not supporting probabilistic matches between strings.)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ipeaGIT/enderecopadrao}
  \item \url{https://ipeagit.github.io/enderecopadrao/}
  \item Report bugs at \url{https://github.com/ipeaGIT/enderecopadrao/issues}
}

}
\author{
\strong{Maintainer}: Daniel Herszenhut \email{dhersz@gmail.com} (\href{https://orcid.org/0000-0001-8066-1105}{ORCID})

}
\keyword{internal}
