% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padronizar_ceps.R
\name{padronizar_ceps}
\alias{padronizar_ceps}
\title{Padronizar CEPs}
\usage{
padronizar_ceps(ceps)
}
\arguments{
\item{ceps}{Um vetor de caracteres ou números. Os CEPs a serem padronizados.}
}
\value{
Um vetor de caracteres com os CEPs padronizados.
}
\description{
Padroniza CEPs.
}
\examples{
ceps <- c("22290-140", "22.290-140", "22290 140", "22290140")
padronizar_ceps(ceps)

ceps <- c(22290140, 1000000, NA)
padronizar_ceps(ceps)

}
