% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padronizar_enderecos.R
\name{padronizar_enderecos}
\alias{padronizar_enderecos}
\title{Padronizar endereços}
\usage{
padronizar_enderecos(enderecos, campos_do_endereco = correspondencia_campos())
}
\arguments{
\item{enderecos}{Um dataframe. Os endereços a serem padronizados. Cada uma de
suas colunas deve corresponder a um campo do endereço (e.g. logradouro,
cidade, bairro, etc).}

\item{campos_do_endereco}{Um vetor nomeado de caracteres. A correspondência
entre os campos a serem padronizados (nomes do vetor) e as colunas que os
representam no dataframe (valores em si). A função
\code{correspondencia_campos()} facilita a criação deste vetor, fazendo também
algumas verificações do conteúdo imputado. Argumentos dessa função com
valor \code{NULL} são ignorados, e ao menos um valor diferente de nulo deve ser
fornecido. Caso deseje criar o vetor manualmente, note que seus nomes devem
ser os mesmos nomes dos parâmetros da função \code{correspondencia_campos()}.}
}
\value{
Um dataframe com os campos de endereço padronizados.
}
\description{
Padroniza um dataframe contendo diversos campos de um endereço.
}
\examples{
path <- file.path(
  Sys.getenv("RESTRICTED_DATA_PATH"),
  "B_CADASTRO/CPF/20230816_cpf.csv"
)
cpf <- data.table::fread(path, nrows = 1000)

campos <- correspondencia_campos(
  logradouro = "logradouro",
  numero = "nroLogradouro",
  complemento = "complemento",
  cep = "cep",
  bairro = "bairro",
  municipio = "codmun_dom",
  estado = "uf_dom"
)

enderecos_padronizados <- padronizar_enderecos(cpf, campos)

}
