% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padronizar_estados.R
\name{padronizar_estados}
\alias{padronizar_estados}
\title{Padronizar estados}
\usage{
padronizar_estados(estados)
}
\arguments{
\item{estados}{Um vetor de caracteres ou números. Os estados a serem
padronizados.}
}
\value{
Um vetor de caracteres com os estados padronizados.
}
\description{
Padroniza um vetor de estados. Veja a seção \emph{Detalhes} para mais informações
sobre a padronização.
}
\section{Detalhes}{

TODO: revisar isso aqui
Operações realizadas durante a padronização:
\enumerate{
\item conversão para caracter, se o input for numérico;
\item remoção de espaços em branco antes e depois dos números e de espaços em
branco em excesso entre números;
\item remoção de zeros à esquerda;
\item substituição de números vazios e de variações de SN (SN, S N, S.N., S./N.,
etc) por S/N.
}
}

\examples{
numeros <- c("0210", "001", "1", "", "S N", "S/N", "SN", "0180  0181")
padronizar_numeros(numeros)

numeros <- c(210, 1, 10000, NA)
padronizar_numeros(numeros)

}
