% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padronizar_logradouros_com_tipo.R
\name{padronizar_logradouros_com_tipo}
\alias{padronizar_logradouros_com_tipo}
\title{Padronizar logradouros e tipos de logradouro simultaneamente}
\usage{
padronizar_logradouros_com_tipo(
  enderecos,
  campos_do_logradouro = correspondencia_logradouro(),
  manter_cols_extras = TRUE
)
}
\arguments{
\item{enderecos}{Um dataframe. Os endereços a serem padronizados, deve
incluir uma coluna com o tipo de logradouro e outra com o logradouro em si.}

\item{campos_do_logradouro}{Um vetor nomeado de caracteres. A correspondência
entre os campos a serem padronizados (nomes do vetor) e as colunas que os
representam no dataframe (valores em si). a função
\code{correspondencia_logradouro()} facilita a criação deste vetor, fazendo
também algumas verificações do conteúdo imputado. caso deseje criar o vetor
manualmente, note que seus nomes devem ser \code{logradouros} e
\code{tipos_de_logradouro}.}

\item{manter_cols_extras}{Um logical. Se colunas não especificadas em
\code{campos_do_logradouro} devem ser mantidas no output ou não (por exemplo,
uma coluna com a informação de bairro ou de id do conjunto de dados sendo
padronizado). Por padrão, \code{TRUE}.}
}
\value{
Caso \code{manter_cols_extras} seja \code{TRUE}, o mesmo dataframe de input,
mas sem as colunas descrevendo o logradouro e o tipo de logradouro e com
uma coluna padronizada adicional \code{logradouro_com_tipo}. Caso
\code{manter_cols_extras} seja \code{FALSE}, um dataframe de apenas uma coluna,
\code{logradouro_com_tipo}.
}
\description{
Padroniza o logradouro e seu tipo (dois campos diferentes) de forma
simultânea, produzindo um novo campo (\code{logradouro_com_tipo}) que garante a
consistência dos dados.
}
\examples{
enderecos <- data.frame(
  id = 1,
  tipo_logradouro = "r",
  logradouro = "ns sra da piedade",
  nroLogradouro = 20,
  complemento = "qd 20",
  cep = 25220020,
  bairro = "jd botanico",
  codmun_dom = 3304557,
  uf_dom = "rj"
)

campos <- correspondencia_logradouro(
  tipo_de_logradouro = "tipo_logradouro",
  logradouro = "logradouro"
)

padronizar_logradouros_com_tipo(enderecos, campos)

padronizar_logradouros_com_tipo(
  enderecos,
  campos,
  manter_cols_extras = FALSE
)

}
