% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correspondencia_campos.R
\name{correspondencia_campos}
\alias{correspondencia_campos}
\title{Correspondência entre campos e colunas}
\usage{
correspondencia_campos(
  tipo_de_logradouro = NULL,
  logradouro = NULL,
  numero = NULL,
  complemento = NULL,
  cep = NULL,
  bairro = NULL,
  municipio = NULL,
  estado = NULL
)
}
\arguments{
\item{tipo_de_logradouro, logradouro, numero, complemento, cep, bairro, municipio, estado}{Uma string. O nome da coluna que representa o respectivo campo de endereço
no dataframe. Pode ser \code{NULL}, no caso do campo não estar listado. Ao menos
um dos campos deve receber um valor não nulo.}
}
\value{
Um vetor nomeado de caracteres, em que os nomes representam os campos
do endereço e os valores as colunas que os descrevem no dataframe.
}
\description{
Cria um vetor de caracteres que especifica as colunas que representam cada
campo de endereço em um dataframe.
}
\examples{
correspondencia_campos(
  tipo_de_logradouro = "tipo_de_logradouro",
  logradouro = "logradouro",
  numero = "nroLogradouro",
  complemento = "complemento",
  cep = "cep",
  bairro = "bairro",
  municipio = "codmun_dom",
  estado = "uf_dom"
)

}
