% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correspondencia_logradouro.R
\name{correspondencia_logradouro}
\alias{correspondencia_logradouro}
\title{Correspondência entre os campos de logradouro e tipo de logradouro e suas
colunas}
\usage{
correspondencia_logradouro(tipo_de_logradouro, logradouro)
}
\arguments{
\item{tipo_de_logradouro}{O nome da coluna que representa o tipo de
logradouro no dataframe de endereços.}

\item{logradouro}{O nome da coluna que representa o logradouro no dataframe
de endereços.}
}
\value{
Um vetor nomeado de caracteres, cujos nomes são \code{tipo_de_logradouro}
e \code{logradouro} e os valores as respectivas colunas que os descrevem no
dataframe de endereços.
}
\description{
Cria um vetor de caracteres que especifica as colunas que representam os
campos de logradouro e tipo de logradouro em um dataframe.
}
\examples{
correspondencia_logradouro(
  tipo_de_logradouro = "tipo_de_logradouro",
  logradouro = "logradouro"
)

}
