% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enderecobr.R
\docType{package}
\name{enderecobr}
\alias{enderecobr}
\alias{enderecobr-package}
\alias{pacote-enderecobr}
\title{enderecobr: Padronizador de Endereços Brasileiros (Brazilian Addresses Standardizer)}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Padroniza endereços brasileiros a partir de diferentes critérios. Os métodos de padronização incluem apenas manipulações básicas de strings, não oferecendo suporte a correspondências probabilísticas entre strings. (Standardizes brazilian addresses using different criteria. Standardization methods include only basic string manipulation, not supporting probabilistic matches between strings.)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ipea/enderecobr}
  \item \url{https://ipea.github.io/enderecobr/}
  \item Report bugs at \url{https://github.com/ipea/enderecobr/issues}
}

}
\author{
\strong{Maintainer}: Rafael H. M. Pereira \email{rafa.pereira.br@gmail.com} (\href{https://orcid.org/0000-0003-2125-7465}{ORCID})

Authors:
\itemize{
  \item Rafael H. M. Pereira \email{rafa.pereira.br@gmail.com} (\href{https://orcid.org/0000-0003-2125-7465}{ORCID})
  \item Daniel Herszenhut \email{dhersz@gmail.com} (\href{https://orcid.org/0000-0001-8066-1105}{ORCID})
  \item Gabriel Garcia de Almeida (\href{https://orcid.org/0009-0003-3557-7328}{ORCID})
  \item Lucas Mation (\href{https://orcid.org/0000-0002-7461-932X}{ORCID})
}

}
\keyword{internal}
