% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padronizar_enderecos.R
\name{padronizar_enderecos}
\alias{padronizar_enderecos}
\title{Padronizar endereços}
\usage{
padronizar_enderecos(
  enderecos,
  campos_do_endereco = correspondencia_campos(),
  manter_cols_extras = TRUE
)
}
\arguments{
\item{enderecos}{Um dataframe. Os endereços a serem padronizados. Cada uma de
suas colunas deve corresponder a um campo do endereço (e.g. logradouro,
cidade, bairro, etc).}

\item{campos_do_endereco}{Um vetor nomeado de caracteres. A correspondência
entre os campos a serem padronizados (nomes do vetor) e as colunas que os
representam no dataframe (valores em si). A função
\code{correspondencia_campos()} facilita a criação deste vetor, fazendo também
algumas verificações do conteúdo imputado. Argumentos dessa função com
valor \code{NULL} são ignorados, e ao menos um valor diferente de nulo deve ser
fornecido. Caso deseje criar o vetor manualmente, note que seus nomes devem
ser os mesmos nomes dos parâmetros da função \code{correspondencia_campos()}.}

\item{manter_cols_extras}{Um logical. Se colunas não especificadas em
\code{campos_do_endereco} devem ser mantidas ou não (por exemplo, uma coluna de
id do conjunto de dados sendo padronizado). Por padrão, \code{TRUE}.}
}
\value{
Um dataframe com os campos de endereço padronizados.
}
\description{
Padroniza um dataframe contendo diversos campos de um endereço.
}
\examples{
enderecos <- data.frame(
  id = 1,
  logradouro = "r ns sra da piedade",
  nroLogradouro = 20,
  complemento = "qd 20",
  cep = 25220020,
  bairro = "jd botanico",
  codmun_dom = 3304557,
  uf_dom = "rj"
)

campos <- correspondencia_campos(
  logradouro = "logradouro",
  numero = "nroLogradouro",
  complemento = "complemento",
  cep = "cep",
  bairro = "bairro",
  municipio = "codmun_dom",
  estado = "uf_dom"
)

padronizar_enderecos(enderecos, campos)

padronizar_enderecos(enderecos, campos, manter_cols_extras = FALSE)

}
